% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{FizzBuzz function in R.}
\usage{
fizzbuzz(n)
}
\arguments{
\item{n}{Positive Number > 0 that is the
top of the range of numbers we want to run fizzbuzz on}
}
\value{
A vector of length n with the corresponding fizzbuzz
combinations for each number i within 1 to n.
}
\description{
For every number from 1 to n
}
\details{
This function will output Fizz if the current number
is divisible by 3, Buzz if the current number is
divisible by 5,  and FizzBuzz if the current number
is divisible by both 3 and 5. If none of these are
true for the current number i, output i for that
iteration.
}
